Getting OSGi JAAS to work with Jackrabbit. We can work with existing JR logic by doing some hacking in DelegatingLoginModule.
The DelegatingLoginModule woul check if Configuration can be obtained via SPI. if yes then it would use that and create a LoginContext and perform *proper* JAAS authentication. Otherwise it would simulate JAAS auth by creating LoginModule (specified va the delegateLoginModuleClass property) and use that similar to localAuthContext class in JR. This would ensure that repository can work untill OSGi based JAAS support is not ready and when it becomes ready logic swicthes to that

With this change removed earlier logic of RepositoryStarter to control repository startup order.